Fixed titles in backedn dashboard button sets in chrome, also some CSS improvements for post-instalation dashboard appearance

+++ admin/views/flexicontent/tmpl/default_postinstall.php
@@ -20,7 +20,7 @@ defined( '_JEXEC' ) or die( 'Restricted access' );
 ?>
 <script type="text/javascript">
 	window.addEvent('domready', function() {
-		var ajaxloader = '<img src="components/com_flexicontent/assets/images/ajax-loader.gif" align="center">';
+		var ajaxloader = '<span class="ajax-loader"></span>';
 <?php if(!$this->existfields) : //@TODO must write a class for that!!! I'm a dirty lazy pig :-) ?>
 	$('existfields').addEvent('click', function(e) {
 		var url = "index.php?option=com_flexicontent&task=createdefaultfields&format=raw&<?php echo JSession::getFormToken();?>=1&tmpl=component";
@@ -359,90 +359,107 @@ $('missingversion').addEvent('click', function(e) {
 <?php endif; ?>
 	});
 </script>
-<table class="adminlist" cellspacing="0" cellpadding="0" border="0">
+<table class="adminlist postinstall-tbl" style="margin: 10px 0 10px 10px; border: 0 none;">
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_PUBLISH_ALL_PLUGINS' ); ?>
 		</td>
-		<td id="publishplugins-log">
-			<?php echo $this->allplgpublish ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="publishplugins" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="publishplugins-log" class="install-task">
+				<?php echo $this->allplgpublish ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="publishplugins" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_INSTALL_DEFAULT_TYPE' ); ?>
 		</td>
-		<td id="existtype-log">
-			<?php echo $this->existtype ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existtype" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existtype-log" class="install-task">
+				<?php echo $this->existtype ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existtype" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'Add/update default Menu Item for URLs' ); ?>
 		</td>
-		<td id="existmenuitems-log">
-			<?php echo $this->existmenuitems ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existmenuitems" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existmenuitems-log" class="install-task">
+				<?php echo $this->existmenuitems ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existmenuitems" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_INSTALL_DEFAULT_FIELDS' ); ?>
 		</td>
-		<td id="existfields-log">
-			<?php echo $this->existfields ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existfields" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existfields-log" class="install-task">
+				<?php echo $this->existfields ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existfields" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_INSTALL_MCATS_RELATIONS' ); ?>
 		</td>
-		<td id="existcats-log">
-			<?php echo $this->existcats ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existcats" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existcats-log" class="install-task">
+				<?php echo $this->existcats ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existcats" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_INSTALL_MULTILINGUAL_SUPPORT' ); ?>
 		</td>
-		<td id="existlang-log">
-			<?php echo $this->existlang ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existlang" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existlang-log" class="install-task">
+				<?php echo $this->existlang ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existlang" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_CREATE_DB_INDEXES' ); ?>
 			<?php
-				if (!$this->existdbindexes) {
+				if (!$this->existdbindexes && !empty($this->missingindexes)) {
 					echo "<br/><span class='fc-mssg-inline fc-mssg fc-note'>this may take a long time on big web-sites, if it timeouts (or takes >2 min) then please refresh, and click to create remaining indexes</span>";
 					echo "<br># tables: ". count($this->missingindexes) ." : ";
-					foreach($this->missingindexes as $tblname => $indexes) {
-						if ( isset($indexes['__indexing_started__']) ) {
-							echo "<br/><b>" .$tblname. "</b> (<small style='color:green'>Indexing started</small>)";
-						} else {
-							echo "<br/><b>" .$tblname. "</b> (". count($indexes) ." <small>indexes missing</small>)";
-						}
+					foreach($this->missingindexes as $tblname => $indexes)
+					{
+						echo isset($indexes['__indexing_started__']) ?
+							"<br/><b>" .$tblname. "</b> (<small style='color:green'>Indexing started</small>)" :
+							"<br/><b>" .$tblname. "</b> (". count($indexes) ." <small>indexes missing</small>)";
 					}
 				}
 			?>
 		</td>
-		<td id="existdbindexes-log">
-			<?php echo $this->existdbindexes ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existdbindexes" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existdbindexes-log" class="install-task">
+				<?php echo $this->existdbindexes ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existdbindexes" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_INSTALL_VERSIONS_TABLE' ); ?>
 		</td>
-		<td id="existversions-log">
-			<?php echo $this->existversions ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existversions" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existversions-log" class="install-task">
+				<?php echo $this->existversions ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existversions" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_UPDATE_VERSIONS_DATA' ); ?>
 		</td>
-		<td id="existversionsdata-log">
-			<?php echo $this->existversionsdata ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existversionsdata" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="existversionsdata-log" class="install-task">
+				<?php echo $this->existversionsdata ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existversionsdata" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
@@ -450,55 +467,67 @@ $('missingversion').addEvent('click', function(e) {
 			<?php echo JText::_( 'FLEXI_INSTALL_AUTHORS_TABLE' ); ?>
 		</td>
 		<td id="existauthors-log">
-			<?php echo $this->existauthors ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="existauthors" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			<div class="install-task"><?php echo $this->existauthors ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="existauthors" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?></div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_UPDATE_TEMPORARY_ITEM_DATA' ); ?>
 		</td>
-		<td id="itemcountingdok-log">
-			<?php echo $this->itemcountingdok ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="itemcountingdok" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+		<td>
+			<div id="itemcountingdok-log" class="install-task">
+				<?php echo $this->itemcountingdok ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="itemcountingdok" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>' ; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_SET_PHPTHUMB_CACHE_PERMISSIONS' ); ?>
 		</td>
-		<td id="cachethumb-log">
-			<?php echo $this->cachethumb ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="cachethumb" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+		<td>
+			<div id="cachethumb-log" class="install-task">
+				<?php echo $this->cachethumb ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="cachethumb" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_CLEANUP_TEMPLATE_FILES' ); ?>
 		</td>
-		<td id="deprecatedfiles-log">
-			<?php echo $this->deprecatedfiles ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="deprecatedfiles" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+		<td>
+			<div id="deprecatedfiles-log" class="install-task">
+				<?php echo $this->deprecatedfiles ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="deprecatedfiles" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_CLEANUP_TABLES' ); ?>
 		</td>
-		<td id="oldfieldsdata-log">
-			<?php echo $this->nooldfieldsdata ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="oldfieldsdata" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+		<td>
+			<div id="oldfieldsdata-log" class="install-task">
+				<?php echo $this->nooldfieldsdata ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="oldfieldsdata" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_CURRENT_VERSIONS' ); ?>
 		</td>
-		<td id="missingversion-log">
-			<?php echo $this->missingversion ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="missingversion" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+		<td>
+			<div id="missingversion-log" class="install-task">
+				<?php echo $this->missingversion ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="missingversion" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+			</div>
 		</td>
 	</tr>
 	<tr>
 		<td class="key">
 			<?php echo JText::_( 'FLEXI_INITIAL_PERMISSION' ); ?>
 		</td>
-		<td id="initialpermission-log">
-			<?php echo $this->initialpermission ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span class="button-add"><a id="initialpermission" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+		<td>
+			<div id="initialpermission-log" class="install-task">
+				<?php echo $this->initialpermission ? '<span class="install-ok"></span>' : '<span class="install-notok"></span><span><a class="fc_button fc_simple" id="initialpermission" href="javascript:;">'.JText::_( 'FLEXI_UPDATE' ).'</a></span>'; ?>
+			</div>
 		</td>
 	</tr>
 </table>
